Removing hooks and moving logic to context
